# [chore] [internal/splunk] Don't use pointer for Event.Time field value