--- a/internal/splunk/common.go
+++ b/internal/splunk/common.go
@@ -49,7 +49,7 @@ type AccessTokenPassthroughConfig struct {
 
 // Event represents a metric in Splunk HEC format
 type Event struct {
-	Time       *float64               `json:"time,omitempty"`       // optional epoch time - set to nil if the event timestamp is missing or unknown
+	Time       float64                `json:"time,omitempty"`       // optional epoch time - set to zero if the event timestamp is missing or unknown (will be added at indexing time)
 	Host       string                 `json:"host"`                 // hostname
 	Source     string                 `json:"source,omitempty"`     // optional description of the source of the event; typically the app's name
 	SourceType string                 `json:"sourcetype,omitempty"` // optional name of a Splunk parsing configuration; this is usually inferred by Splunk
@@ -99,14 +99,14 @@ func (e *Event) UnmarshalJSON(b []byte) error {
 	}
 	switch t := rawEvent.Time.(type) {
 	case float64:
-		e.Time = &t
+		e.Time = t
 	case string:
 		{
 			time, err := strconv.ParseFloat(t, 64)
 			if err != nil {
 				return err
 			}
-			e.Time = &time
+			e.Time = time
 		}
 	}
 	return nil