The pointer was added in https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/1157 to avoid sending zero timestamps by Splunk HEC exporter. But using a pointer is not required for that, `json:omitempty` tag does the same for float64 zero values.

After this change, we cannot differentiate `nil`/`0` values of `Time` field in the internal `splunk.Event` struct, but it doesn't matter because both versions produce the same JSON string without `time` field and translates to/from the same OTLP `0` values.

Before:

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
Benchmark_pushLogData_10_10_1024
Benchmark_pushLogData_10_10_1024-10    	   36219	     32740 ns/op	   25431 B/op	     416 allocs/op
Benchmark_pushLogData_10_10_8K
Benchmark_pushLogData_10_10_8K-10      	   13060	     94597 ns/op	   71809 B/op	    1167 allocs/op
Benchmark_pushLogData_10_10_2M
Benchmark_pushLogData_10_10_2M-10      	   12934	     91213 ns/op	   74024 B/op	    1199 allocs/op
Benchmark_pushLogData_10_200_2M
Benchmark_pushLogData_10_200_2M-10     	     634	   1907038 ns/op	 1476323 B/op	   24000 allocs/op
Benchmark_pushLogData_100_200_2M
Benchmark_pushLogData_100_200_2M-10    	      76	  13599245 ns/op	11354600 B/op	  183258 allocs/op
Benchmark_pushLogData_100_200_5M
Benchmark_pushLogData_100_200_5M-10    	      57	  18378855 ns/op	14868081 B/op	  239911 allocs/op
BenchmarkConsumeLogsRejected
BenchmarkConsumeLogsRejected-10        	    1245	    972941 ns/op	  739034 B/op	   12002 allocs/op
PASS
```

After:

```
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
Benchmark_pushLogData_10_10_1024
Benchmark_pushLogData_10_10_1024-10    	   38847	     31386 ns/op	   25167 B/op	     383 allocs/op
Benchmark_pushLogData_10_10_8K
Benchmark_pushLogData_10_10_8K-10      	   13942	     87630 ns/op	   71106 B/op	    1076 allocs/op
Benchmark_pushLogData_10_10_2M
Benchmark_pushLogData_10_10_2M-10      	   13096	     90047 ns/op	   73338 B/op	    1109 allocs/op
Benchmark_pushLogData_10_200_2M
Benchmark_pushLogData_10_200_2M-10     	     631	   1917114 ns/op	 1460391 B/op	   22010 allocs/op
Benchmark_pushLogData_100_200_2M
Benchmark_pushLogData_100_200_2M-10    	      79	  12951741 ns/op	11175612 B/op	  168047 allocs/op
Benchmark_pushLogData_100_200_5M
Benchmark_pushLogData_100_200_5M-10    	      61	  16757523 ns/op	14699264 B/op	  220011 allocs/op
BenchmarkConsumeLogsRejected
BenchmarkConsumeLogsRejected-10        	    1274	    935309 ns/op	  731880 B/op	   11012 allocs/op
PASS
```